+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[Unreleased]
+Added
+Apple Sign In enhancements to get user profile info [SDKS-1632]
+
 # Version 3.2.0
 ## [Unreleased]
 #### Added
@@ -5,6 +9,7 @@
 #### Changed
 - Remove "Accept: application/x-www-form-urlencoded" header from /authorize endpoint for GET requests [SDKS-1729]
 - Fix issue when expired push notification displayed as "Approved" in the notification history list [SDKS-1491]
+- Fix Issues with registering TOTP accounts with invalid period [SDKS-1405]
 
 ## [3.2.0]
 #### Changed